Add test types displayed in CI (#11770)
This is a follow up after #11659. Thanks to this one, it will
be immediately visible in CI which types of tests are run.

This change has been lost during one of the rebases, so it is
brought back now.

(cherry picked from commit cca5d63186bc09def05a9180003dce9d17172640)